+++ src/App.css
@@ -13,18 +13,34 @@ h1, h2, h4, h5 {
 
 h1 {
   display: flex;
-  -webkit-text-stroke:2px rgba(0, 0, 0, 1);
-  font-size: 300px;
-  font-size: 20vw;
   font-weight: 300;
-  color: rgba(255, 255, 255, 0.2);
+  font-size: 300px;
+  font-size: 19vw;
+  line-height: 116.667%;
+  -webkit-text-stroke:2px rgba(0, 0, 0, 0.5);
+  background: radial-gradient(94.84% 94.84% at 50% 5.16%, #90F2E1 0%, #FFF 100%);
+  mix-blend-mode: luminosity;
   -webkit-background-clip: text;
-  padding: 0 50px;
-  margin: 0;
+  -webkit-text-fill-color: transparent;
   z-index: 1;
+}
 
 
-
+h3 {
+  font-family: 'Just Me Again Down Here', cursive;
+  font-weight: 400;
+  font-style: normal;
+  line-height: normal;
+  font-size: 95px;
+  text-align: center;
+  -webkit-text-stroke:2px rgba(0, 0, 0, 0.5);
+  opacity: 0.75;
+  background: conic-gradient(from 180deg at 50% 50.00%, #F0EDFA 0deg, #E8E4F9 290.6250071525574deg);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  mix-blend-mode: luminosity;
+  z-index: 1;
 }
 
 h2 {
@@ -39,30 +55,8 @@ h2 {
   /* identical to box height */
   text-align: center;
   color: #FFFFFF;
-
-
 }
 
-h3 {
-  font-family: 'Just Me Again Down Here', cursive;
-  font-weight: 400;
-  font-size: 100px;
-  text-align: center;
-  -webkit-text-stroke:2px rgba(0, 0, 0, 1);
-  background: conic-gradient(from 180deg at 50% 50%, #F0EDFA 0deg, #E8E4F9 290.63deg, #F0EDFA 360deg);
-  opacity: 0.65;
-  mix-blend-mode: luminosity;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  z-index: 1;
-}
-
-h4 {
-  color: rgba(255, 255, 255, 0.50);
-  text-align: center;
-  font-size: 65px;
-  line-height: 24px;
-}
 
 h5 {
   font-size: 65px;
@@ -70,17 +64,24 @@ h5 {
 
 }
 
-
 p {
   font-family: "italiana", sans-serif;
   margin: 12px 0;
   color: #FFFFFF;
 }
 
+
 .MacBook-Pro-14 {
+  width: 1293px;
+  height: 885px;
   display: flex;
   background: #90F2E1;
   background:linear-gradient(70deg,rgba(239, 175, 128,0.51),rgba(144,242,225,1));
+  /*background: linear-gradient(207deg, #90F2E1 0%, rgba(239, 175, 128, 0.31) 100%);*/
+  background-repeat: no-repeat;
+  margin: 0;
+  background-attachment: fixed;
+
 }
 
 .outer-content-container {
@@ -97,7 +98,7 @@ p {
   flex-grow: 0;
   /*margin: 0 20px;*/
   /*display: flex;*/
-  /*flex-direction: column;*/
+  /*flex-direction: column;
   /*align-items: center;*/
 
 }
@@ -111,14 +112,12 @@ p {
   padding: 100px 0;
 }
 
-button {
-  font-family: 'Inter', sans-serif;
-  text-align: center;
+
+.signup-button {
+  font-family: Open Sans, sans-serif;
+  font-size: 85px;
   font-style: normal;
   font-weight: 600;
-  font-size: 32px;
-  cursor: pointer;
-  color: #FFFFFF;
+  line-height: 16px; /* 44.444% */
 
 }
-
