+++ src/Context/AuthContext.js
@@ -37,7 +37,7 @@ function AuthContextProvider({children}) {
                 }, status:'done'
             })
             console.log(result.data)
-            history.push('/profile');
+            history.push('/welcomepage');
         } catch (error) {
             console.error();
             toggleAuth({...auth, isAuth: false,
