+++ src/Pages/Homepage/Homepage.js
@@ -16,58 +16,56 @@ function Homepage() {
 
 
     return (
-        <main className={styles["outer-content-container"]}>
+       <>
             <section>
                     <div className={styles["rectangles-headers-container"]}>
-                            <h3 className={styles["easy-way-to-move"]}>
-                                The easy way to move
-                            </h3>
-                        <div className={styles["rectangle-small-container"]}>
-                            <img className={styles["rectangle-small"]} src={waters} alt="lilac waters"/>
-                        </div>
                         <div className={styles["yoga-space-container"]}>
                             <h1 className={styles["yoga"]}>
                                 Yoga
                             </h1>
                             <h1 className={styles["space"]}>
-                                Space
+                                Spac<div className={styles["e"]}>e</div>
                             </h1>
-                        </div>
-                        <div className={styles["rectangle-big-container"]}>
-                            <img className={styles["rectangle-big"]} src={waters} alt="lilac waters"/>
-                        </div>
 
+                            <div className={styles["rectangle-big-container"]}>
+                                <img className={styles["rectangle-big"]} src={waters} alt="lilac waters"/>
+                            </div>
+                        </div>
+                            <div className={styles["easy-way-to-move-container"]}>
+                            <h3 className={styles["easy-way-to-move"]}>
+                                The easy way to move
+                            </h3>
+                            </div>
+                        <div className={styles["rectangle-small-container"]}>
+                            <img className={styles["rectangle-small"]} src={waters} alt="lilac waters"/>
+                        </div>
                     </div>
             </section>
-
-
-                    <section>
-
-                        <div className={styles["inner-content-container sign-up-or-in"]}>
-
-                            <div className={styles["register-container"]}>
-                            <button
-                                type="button"
-                                onClick={handleClickRegister}
-                                className={styles["register-button"]}>
-                                Register
-                            </button>
-                            </div>
-
-                        <p className={styles["under-register-text"]}>For daily yoga video's</p>
-
-                        <div className={styles["sign-in-container"]}>
-                            <button
-                                type="button"
-                                onClick={handleClickSignin}
-                                className={styles["sign-in-button"]}>
-                                Sign in
-                            </button>
-                        </div>
-                        <p className={styles["under-sign-in-text"]}>To enter your yoga space</p>
-                        </div>
-                    </section>
-                </main>
+           <section>
+               <div className={styles["inner-content-container sign-up-or-in"]}>
+                   <div className={styles["register-container"]}>
+                       <button
+                           type="button"
+                           onClick={handleClickRegister}
+                           className={styles["register-button"]}>
+                           Register
+                       </button>
+                   </div>
+                   <div className={styles["sign-in-container"]}>
+                       <button
+                           type="button"
+                           onClick={handleClickSignin}
+                           className={styles["sign-in-button"]}>
+                           Sign in
+                       </button>
+                   </div>
+                   {/*<p className={styles["under-sign-in-text"]}>To enter your yoga space</p>*/}
+               </div>
+           </section>
+           {/*<div className={styles["register-text-container"]}>*/}
+           {/*    <p className={styles["under-register-text"]}>For daily yoga video's</p>*/}
+           {/*</div>*/}
+           </>
 
             );
 }
