+++ src/Pages/Homepage/Homepage.module.css
@@ -1,95 +1,192 @@
-.rectangles-headers-container {
+.yoga-space-container {
+
     display: flex;
+    text-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
+    width: 1194px;
+    height: 738px;
     flex-direction: column;
-    height: calc(100vh - 600px);
-    border: 3px dashed purple;
+    flex-shrink: 0;
+
+}
+
+.yoga {
+
+    text-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
+    opacity: 38%;
+
+}
+
+.space {
+    text-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
+    opacity: 38%;
+}
+
+.e {
+    color: #FFFFFF;
+    text-size-adjust: inherit;
+    text-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
+    -webkit-text-fill-color: #FFFFFF;
+}
 
 
+.easy-way-to-move-container {
+    display: flex;
+    width: 582px;
+    height: 272px;
+    flex-direction: column;
+    justify-content: flex-end;
+    position: absolute;
+    top: 580px;
+    right: 90px;
 }
 
 
 .easy-way-to-move {
-    justify-content: flex-end;
-    position: absolute;
-    top: 500px;
-    width: 500px;
-    right: 50px;
-    word-wrap: break-word;
-    margin-right: 80px;
-    z-index: 1;
+    text-shadow: 4px 4px 0px rgba(0, 0, 0, 0.25);
 }
 
+
 .rectangle-small-container {
+
     display: flex;
     /*height: calc(100vh - 150px);*/
     justify-content: flex-end;
-    position: relative;
-    border: 3px dashed greenyellow;
+    position: absolute;
+    top:750px;
+    right:90px;
+    /*border: 3px dashed greenyellow;*/
+
 }
+
+
 
 .rectangle-small {
+
+    width: 1179px;
+    height: 153px;
     position: relative;
-    top:700px;
-    right:180px;
-    width: 700px;
-    height: 100px;
-    margin: 0 25px 50px 0;
     z-index: 0;
+    flex-shrink: 0;
+
 }
 
+/*.sign-up-or-in {*/
+/*    position: sticky;*/
+/*}*/
 
 
-.yoga-space-container {
-    display: flex;
-    /*flex-direction: row;*/
-    flex-wrap: wrap;
-    column-gap: 16px;
-    /*height: calc(100vh - 770px);*/
-    border: 3px dashed blue;
-}
+.rectangle-big-container {
 
 
-.yoga {
-    align-self: flex-start;
-    margin-left: 20px;
-    height: 380px;
-    z-index: 1;
-    flex-grow: 0;
-    flex-shrink: 1;
-}
-
-.space {
-    align-self: flex-start;
-    flex-wrap: wrap;
-    height: 390px;
-    z-index: 1;
-}
-
-.rectangle-big-container{
     display: flex;
     justify-content: center;
-    height: calc(100vh - 220px);
-    border: 3px dashed orangered;
+    position: absolute;
+    /*border: 3px dashed orangered;*/
+    bottom: 0;
+    flex-shrink: 0;
 }
 
 .rectangle-big {
-    width: 1000px;
+    margin: 668px 85px 18px 86px;
     width: 90vw;
-    height: 170px;
-    margin: 0 0 50px 0;
+    height: 199px;
+    position: relative;
     z-index: 0;
+    flex-shrink: 0;
 
 }
+
+.register-container {
+    position: absolute;
+    width: 437px;
+    height: 214px;
+    flex-shrink: 0;
+    margin: 321px 734px 110px 122px;
+}
+
 
 .register-button {
-    z-index: 1;
-    align-items: center;
+
+    display: flex;
+    position: relative;
+
     width: 213px;
     height: 48px;
+    padding: 21px 24px;
+    align-items: center;
     border: none;
-    background: rgba(144, 242, 225, 0.31);
+    justify-content: center;
+    flex-shrink: 0;
     border-radius: 8px;
+    background: rgba(144, 242, 225, 0.31);
+    z-index: 1;
+    font-family: 'Inter', sans-serif;
+    text-align: center;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 32px;
+    cursor: pointer;
+    color: #FFFFFF;
 }
+
+
+/*.register-text-container {*/
+/*    position: absolute;*/
+/*    width: 332px;*/
+/*    height: 51px;*/
+/*    bottom:1px;*/
+/*    left:155px;*/
+/*    flex-shrink: 0;*/
+
+/*}*/
+/*.under-register-text {*/
+
+/*    position: relative;*/
+/*    text-align: center;*/
+/*    font-size: 33px;*/
+/*    font-style: normal;*/
+/*    font-weight: 400;*/
+/*    line-height: normal;*/
+/*}*/
+
+.sign-in-container {
+
+    width: 437px;
+    height: 214px;
+    position: absolute;
+    align-content: flex-end;
+    flex-shrink: 0;
+    margin: 289px 0 110px 0;
+    border-radius: 20px;
+}
+
+.sign-in-button {
+
+    display: flex;
+    position: relative;
+    left: 1989px;
+    top: 24px;
+    width: 213px;
+    height: 48px;
+    padding: 21px 24px;
+    align-items: center;
+    border: none;
+    justify-content: center;
+    border-radius: 8px;
+    background: rgba(144, 242, 225, 0.31);
+    flex-shrink: 0;
+    font-family: 'Inter', sans-serif;
+    text-align: center;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 32px;
+    cursor: pointer;
+    color: #FFFFFF;
+
+
+}
+
+
 /*.sign-up-or-in {*/
 /*    display: flex;*/
 /*    flex-direction: row;*/
@@ -153,5 +250,4 @@
 /*.sign-in-button:hover {*/
 /*    color: #EFAF80;*/
 /*    border: 5px solid plum;*/
-
 /*}*/
