+++ src/Pages/Signin/SignIn.module.css
@@ -1,4 +1,18 @@
+.sign-in-header {
+    width: 553px;
+    height: 122px;
+    flex-shrink: 0;
+    color: #FFF;
+    justify-content: center;
+    text-align: center;
+    position: absolute;
+    left: 1000px;
+    text-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
+    font-size: 66px;
+    font-weight: 600;
+    line-height: normal;
 
+}
 
 
 
@@ -43,16 +57,34 @@ h5  {
     left: 25%;
 }
 
-.submit-button  {
-    position:   absolute;
-    left: 18%;
-    top: 95%;
-    transform: translate(-50%, -50%);
-    -ms-transform: translate(-50%, -50%);
-    color: transparent;
-    font-size: 16px;
-    padding: 12px 24px;
-    cursor: pointer;
-    border-radius: 86px;
-    border-color: transparent;
+.sign-in-submit-button  {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 527px;
+    height: 82px;
+    margin: 96px 5px;
+    color: #FFF;
+    text-align: center;
+    font-size: 36px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 16px; /* 44.444% */
+    flex-shrink: 0;
+    border: 2px solid #90F2E1;
+    border-radius: 8px;
+    background: linear-gradient(180deg, #90F2E1 0%, rgba(239, 175, 128, 0.31) 100%);
+
+    /*position:   absolute;*/
+    /*left: 18%;*/
+    /*top: 95%;*/
+    /*transform: translate(-50%, -50%);*/
+    /*-ms-transform: translate(-50%, -50%);*/
+    /*color: transparent;*/
+    /*font-size: 16px;*/
+    /*padding: 12px 24px;*/
+    /*cursor: pointer;*/
+    /*border-radius: 86px;*/
+    /*border-color: transparent;*/
 }
